solves 5_B with hammer because my pc cannot read the last 3 in the first line

+++ exercises/ex_5.h
@@ -3,6 +3,6 @@
 
 #include "../core/read_file.h"
 
-long ex_5(array_string *result);
+long long ex_5(array_string *result);
 
 #endif
